draft: add host control
