+++ project-4/common/types/index.ts
@@ -64,4 +64,4 @@ export type MediaSetup = {
 };
 
 export type Status = 'loading' | 'idle' | 'rejected' | 'success';
-
+export type Kind = 'audio' | 'video' | 'chat' | 'screen' | 'fullscreen';
