+++ project-4/components/streams/index.ts
@@ -1,0 +1,5 @@
+import MyStream from './my-stream';
+import OtherStreams from './other-streams';
+import SharedScreenStream from './shared-screen-stream';
+
+export { MyStream, OtherStreams, SharedScreenStream };
