+++ project-4/contexts/users-settings.tsx
@@ -1,0 +1,59 @@
+import { createContext, useContext, useEffect, useState } from 'react';
+import { useRouter } from 'next/router';
+
+import { KeyValue, Nullable, PeerId, RoomId } from '@common/types';
+import { append, isHost } from '@common/utils';
+
+import { SocketContext } from '@pages/_app';
+
+export const UsersUpdaterContext = createContext<any>({});
+export const UsersStateContext = createContext<any>({});
+
+export default function UsersSettingsProvider({ children }: any) {
+  const router = useRouter();
+  const socket = useContext(SocketContext);
+
+  const [streams, setStreams] = useState<Record<PeerId, JSX.Element>>({});
+
+  const [isMuted, setIsMuted] = useState<KeyValue<boolean>>({});
+  const [isHidden, setIsHidden] = useState<KeyValue<boolean>>({});
+  const [avatars, setAvatars] = useState<KeyValue<string>>({});
+
+  const [sharedScreenTrack, setSharedScreenTrack] =
+    useState<Nullable<MediaStreamTrack>>(null);
+
+  useEffect(() => {
+    socket.on('user:toggled-audio', (peerId: PeerId) =>
+      setIsMuted(append({ [peerId]: !isMuted[peerId] }))
+    );
+
+    socket.on('user:toggled-video', (peerId: PeerId) =>
+      setIsHidden(append({ [peerId]: !isHidden[peerId] }))
+    );
+  }, [isMuted, isHidden]);
+
+  return (
+    <UsersStateContext.Provider
+      value={{
+        streams,
+        isMuted,
+        isHidden,
+        isHost: isHost(router.query.qoraId as RoomId),
+        avatars,
+        sharedScreenTrack,
+      }}
+    >
+      <UsersUpdaterContext.Provider
+        value={{
+          setIsMuted,
+          setIsHidden,
+          setAvatars,
+          setStreams,
+          setSharedScreenTrack,
+        }}
+      >
+        {children}
+      </UsersUpdaterContext.Provider>
+    </UsersStateContext.Provider>
+  );
+}
